requirements.txt
python3-discogs-client
